feat: move files to network and logic
