+++ libs/coin-modules/coin-aptos/src/__tests__/bridge/broadcast.test.ts
@@ -1,10 +1,10 @@
 import broadcast from "../../bridge/broadcast";
-import { AptosAPI } from "../../api";
+import { AptosAPI } from "../../network";
 import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
 import type { Account, Operation, SignedOperation } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
 
-jest.mock("../../api");
+jest.mock("../../network");
 jest.mock("@ledgerhq/coin-framework/operation");
 
 describe("broadcast", () => {
